+++ sipXmediaLib/include/mp/StreamHttpDataSource.h
@@ -33,7 +33,7 @@ class OsConnectionSocket ;
 class HttpMessage ;
 // TYPEDEFS
 typedef UtlBoolean (*GetDataCallbackProc)(char* pData, 
-                                         size_t iLength, 
+                                         ssize_t iLength, 
                                          void* pOptionalData, 
                                          HttpMessage* pMsg);
 
@@ -74,7 +74,7 @@ public:
    virtual OsStatus destroyAndDelete() ;
      //:Destroys and deletes the data source object
 
-   virtual OsStatus read(char *szBuffer, size_t iLength, size_t& iLengthRead) ;
+   virtual OsStatus read(char *szBuffer, ssize_t iLength, ssize_t& iLengthRead) ;
      //:Reads iLength bytes of data from the data source and places the
      //:data into the passed szBuffer buffer.
      //
@@ -82,7 +82,7 @@ public:
      //!param iLength - Max length to read
      //!param iLengthRead - The actual amount of data read.
 
-   virtual OsStatus peek(char* szBuffer, size_t iLength, size_t& iLengthRead) ;
+   virtual OsStatus peek(char* szBuffer, ssize_t iLength, ssize_t& iLengthRead) ;
      //:Identical to read, except the stream pointer is not advanced.
      //
      //!param szBuffer - Buffer to place data
@@ -102,16 +102,16 @@ public:
      //!param iLocation - The desired seek location
 
 
-   UtlBoolean deliverData(char *szData, int iLength, size_t iMaxLength) ;
+   UtlBoolean deliverData(char *szData, ssize_t iLength, ssize_t iMaxLength) ;
      //:Callback routine that is invoked whenever new data is available from
      //:http socket.
 
 /* ============================ ACCESSORS ================================= */
 
-   virtual OsStatus getLength(size_t& iLength);      
+   virtual OsStatus getLength(ssize_t& iLength);      
      //:Gets the length of the stream (if available)
 
-   virtual OsStatus getPosition(size_t& iPosition) ;
+   virtual OsStatus getPosition(ssize_t& iPosition) ;
      //:Gets the current position within the stream.
    
    virtual OsStatus toString(UtlString& string) ;
